old bugfix

+++ include/ucui.h
@@ -132,7 +132,7 @@ struct arm_regs * read_arm_registers(uc_engine *uc);
 void printwass(unsigned int startpos, unsigned int endpos, uint64_t pc);
 struct disassembly * disass(uint8_t *code, unsigned int len, uint64_t baseaddress, cs_arch arch, cs_mode mode);
 struct readfile * readfile(char *filename);
-void verify_visible_ip(uint32_t pc);
+void verify_visible_ip(uint64_t pc);
 bool ip_aligned_to_disassembly(uint64_t pc);
 bool should_break(uint64_t pc);
 void redisassemble_code(uc_engine *uc, uint64_t ip);
